5주차 스프린트 기능 반영

+++ backend/src/product/product.service.ts
@@ -11,16 +11,18 @@ import { ProductDetailsDto } from 'src/dto/product.details.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import { ProductPrice } from 'src/schema/product.schema';
 import { Model } from 'mongoose';
-import { ProductPriceDto } from 'src/dto/product.price.dto';
 import { PriceDataDto } from 'src/dto/price.data.dto';
-import { KR_OFFSET, NINETY_DAYS, NO_CACHE, THIRTY_DAYS } from 'src/constants';
-import { Cron } from '@nestjs/schedule';
+import { MAX_TRACKING_RANK, NINETY_DAYS, NO_CACHE, THIRTY_DAYS } from 'src/constants';
+import { ProductRankCache } from 'src/utils/cache';
+import { ProductRankCacheDto } from 'src/dto/product.rank.cache.dto';
+import Redis from 'ioredis';
+import { InjectRedis } from '@songkeys/nestjs-redis';
 
 const REGEXP_11ST =
     /http[s]?:\/\/(?:www\.|m\.)?11st\.co\.kr\/products\/(?:ma\/|m\/|pa\/)?([1-9]\d*)(?:\?.*)?(?:\/share)?/;
 @Injectable()
 export class ProductService {
-    private productDataCache = new Map();
+    private productRankCache = new ProductRankCache(MAX_TRACKING_RANK);
     constructor(
         @InjectRepository(TrackingProductRepository)
         private trackingProductRepository: TrackingProductRepository,
@@ -28,6 +30,7 @@ export class ProductService {
         private productRepository: ProductRepository,
         @InjectModel(ProductPrice.name)
         private productPriceModel: Model<ProductPrice>,
+        @InjectRedis() private readonly redis: Redis,
     ) {
         this.initCache();
     }
@@ -47,14 +50,31 @@ export class ProductService {
                     },
                 },
             ])
+            .sort('_id')
             .exec();
-        latestData.forEach((data) => {
-            this.productDataCache.set(data._id, {
-                price: data.price,
-                isSoldOut: data.isSoldOut,
-                lowestPrice: data.lowestPrice,
-            });
+        const userCountList = await this.trackingProductRepository.getAllUserCount();
+        const rankList = await this.productRepository.getTotalInfoRankingList();
+        const initPromise = latestData.map(async (data) => {
+            const matchProduct = userCountList.find((product) => product.id === data._id);
+            const setUserCount = await this.redis.set(
+                `product:${data._id}`,
+                JSON.stringify({
+                    isSoldOut: data.isSoldOut,
+                    price: data.price,
+                    lowestPrice: data.lowestPrice,
+                }),
+            );
+            const zaddUserCount = await this.redis.zadd(
+                'userCount',
+                matchProduct ? parseInt(matchProduct.userCount) : 0,
+                data._id,
+            );
+            return Promise.all([setUserCount, zaddUserCount]);
         });
+        rankList.forEach((product) => {
+            this.productRankCache.put(product.id, { ...product, userCount: parseInt(product.userCount) });
+        });
+        await Promise.all(initPromise);
     }
 
     async verifyUrl(productUrlDto: ProductUrlDto): Promise<ProductInfoDto> {
@@ -72,13 +92,26 @@ export class ProductService {
         const existProduct = await this.productRepository.findOne({
             where: { productCode: productCode },
         });
-        const productInfo = existProduct ?? (await getProductInfo11st(productCode));
-        const product = existProduct ?? (await this.productRepository.saveProduct(productInfo));
+        const product = existProduct ?? (await this.firstAddProduct(productCode));
         const trackingProduct = await this.trackingProductRepository.findOne({
             where: { productId: product.id, userId: userId },
         });
         if (trackingProduct) {
             throw new HttpException('이미 등록된 상품입니다.', HttpStatus.CONFLICT);
+        }
+        const cacheData = await this.redis.zscore('userCount', product.id);
+        const userCount = cacheData ? parseInt(cacheData) : NO_CACHE;
+        const productRank = {
+            id: product.id,
+            productName: product.productName,
+            productCode: product.productCode,
+            shop: product.shop,
+            imageUrl: product.imageUrl,
+            userCount: userCount + 1,
+        };
+        this.productRankCache.update(productRank);
+        if (cacheData) {
+            await this.redis.zincrby('userCount', 1, product.id);
         }
         await this.trackingProductRepository.saveTrackingProduct(userId, product.id, targetPrice);
     }
@@ -88,12 +121,13 @@ export class ProductService {
             where: { userId: userId },
             relations: ['product'],
         });
-        if (trackingProductList.length === 0) {
-            throw new HttpException('상품 목록을 찾을 수 없습니다.', HttpStatus.NOT_FOUND);
-        }
-        const trackingListInfo = trackingProductList.map(async ({ product, targetPrice }) => {
+        if (trackingProductList.length === 0) return [];
+        const trackingListInfo = trackingProductList.map(async ({ product, targetPrice, isAlert }) => {
             const { id, productName, productCode, shop, imageUrl } = product;
-            const { price } = this.productDataCache.get(id) ?? { price: NO_CACHE };
+            const cacheData = await this.redis.get(`product:${id}`);
+            const { price } = cacheData
+                ? JSON.parse(cacheData)
+                : await this.productRepository.findOne({ where: { id: id } });
             const priceData = await this.getPriceData(id, THIRTY_DAYS);
             return {
                 productName,
@@ -102,6 +136,7 @@ export class ProductService {
                 imageUrl,
                 targetPrice: targetPrice,
                 price,
+                isAlert,
                 priceData,
             };
         });
@@ -110,10 +145,13 @@ export class ProductService {
     }
 
     async getRecommendList() {
-        const recommendList = await this.trackingProductRepository.getTotalInfoRankingList();
+        const recommendList = this.productRankCache.getAll();
         const recommendListInfo = recommendList.map(async (product, index) => {
             const { id, productName, productCode, shop, imageUrl } = product;
-            const { price } = this.productDataCache.get(id) ?? { price: NO_CACHE };
+            const cacheData = await this.redis.get(`product:${id}`);
+            const { price } = cacheData
+                ? JSON.parse(cacheData)
+                : await this.productRepository.findOne({ where: { id: id } });
             const priceData = await this.getPriceData(id, THIRTY_DAYS);
             return {
                 productName,
@@ -139,11 +177,11 @@ export class ProductService {
         const trackingProduct = await this.trackingProductRepository.findOne({
             where: { userId: userId, productId: selectProduct.id },
         });
-        const ranklist = await this.trackingProductRepository.getRankingList();
-        const idx = ranklist.findIndex(({ id }) => id === selectProduct.id);
+        await this.trackingProductRepository.getUserCount(selectProduct.id);
+        const idx = this.productRankCache.findIndex(selectProduct.id);
         const rank = idx === -1 ? idx : idx + 1;
         const priceData = await this.getPriceData(selectProduct.id, NINETY_DAYS);
-        const { price, lowestPrice } = this.productDataCache.get(selectProduct.id);
+        const { price, lowestPrice } = await this.getProductCurrentData(selectProduct.id);
         return {
             productName: selectProduct.productName,
             shop: selectProduct.shop,
@@ -160,12 +198,49 @@ export class ProductService {
     async updateTargetPrice(userId: string, productAddDto: ProductAddDto) {
         const product = await this.findTrackingProductByCode(userId, productAddDto.productCode);
         product.targetPrice = productAddDto.targetPrice;
+        product.isFirst = true;
         await this.trackingProductRepository.save(product);
     }
 
     async deleteProduct(userId: string, productCode: string) {
         const product = await this.findTrackingProductByCode(userId, productCode);
+        const prevProduct = this.productRankCache.get(product.productId)?.value;
+        await this.redis.zincrby('userCount', -1, product.productId);
+
+        if (!prevProduct) {
+            throw new HttpException('상품을 찾을 수 없습니다.', HttpStatus.NOT_FOUND);
+        }
+        prevProduct.userCount--;
+        const productCount = await this.redis.zcard('userCount');
+        if (productCount > MAX_TRACKING_RANK) {
+            await this.deleteUpdateCache(prevProduct);
+        } else {
+            this.productRankCache.update(prevProduct);
+        }
         await this.trackingProductRepository.remove(product);
+    }
+
+    async deleteUpdateCache(prevProduct: ProductRankCacheDto) {
+        const nextDataId = (await this.redis.zrevrange('userCount', MAX_TRACKING_RANK, MAX_TRACKING_RANK))[0];
+        const cacheData = await this.redis.zscore('userCount', nextDataId);
+        const userCount = cacheData ? parseInt(cacheData) : NO_CACHE;
+        if (userCount >= prevProduct.userCount) {
+            const newProduct = await this.productRepository.findOne({
+                where: { id: nextDataId },
+            });
+            if (!newProduct) {
+                throw new HttpException('상품을 찾을 수 없습니다.', HttpStatus.NOT_FOUND);
+            }
+            const newProductRanck = {
+                id: newProduct.id,
+                productName: newProduct.productName,
+                productCode: newProduct.productCode,
+                shop: newProduct.shop,
+                imageUrl: newProduct.imageUrl,
+                userCount: userCount,
+            };
+            this.productRankCache.update(prevProduct, newProductRanck);
+        }
     }
 
     async findTrackingProductByCode(userId: string, productCode: string) {
@@ -184,13 +259,8 @@ export class ProductService {
         return trackingProduct;
     }
 
-    async mongo(productPriceDto: ProductPriceDto) {
-        const newData = new this.productPriceModel(productPriceDto);
-        return newData.save();
-    }
-
     async getPriceData(productId: string, days: number): Promise<PriceDataDto[]> {
-        const endDate = new Date(+new Date() + KR_OFFSET);
+        const endDate = new Date();
         const startDate = new Date(endDate);
         startDate.setDate(endDate.getDate() - days);
         const dataInfo = await this.productPriceModel
@@ -206,26 +276,58 @@ export class ProductService {
             return { time: new Date(time).getTime(), price, isSoldOut };
         });
     }
-    @Cron('* */10 * * * *')
-    async cyclicPriceChecker() {
-        const productList = await this.productRepository.find({ select: { id: true, productCode: true } });
-        const productCodeList = productList.map(({ productCode, id }) => getProductInfo11st(productCode, id));
-        const results = (await Promise.all(productCodeList)).map(({ productId, productPrice, isSoldOut }) => {
-            return { productId, price: productPrice, isSoldOut };
-        });
-        const updatedDataInfo = results.filter(({ productId, price, isSoldOut }) => {
-            const cache = this.productDataCache.get(productId);
-            if (!cache || cache.isSoldOut !== isSoldOut || cache.price !== price) {
-                const lowestPrice = cache ? Math.min(cache.lowestPrice, price) : price;
-                this.productDataCache.set(productId, {
-                    isSoldOut,
-                    price,
-                    lowestPrice,
-                });
-                return true;
-            }
-            return false;
-        });
-        await this.productPriceModel.insertMany(updatedDataInfo);
+
+    async firstAddProduct(productCode: string) {
+        const productInfo = await getProductInfo11st(productCode);
+        const product = await this.productRepository.saveProduct(productInfo);
+        const updatedDataInfo = {
+            productId: product.id,
+            price: productInfo.productPrice,
+            isSoldOut: productInfo.isSoldOut,
+        };
+        this.redis.set(
+            `product:${product.id}`,
+            JSON.stringify({
+                price: productInfo.productPrice,
+                isSoldOut: productInfo.isSoldOut,
+                lowestPrice: productInfo.productPrice,
+            }),
+        );
+        this.productPriceModel.create(updatedDataInfo);
+        return product;
+    }
+
+    async toggleProductAlert(userId: string, productCode: string) {
+        const product = await this.findTrackingProductByCode(userId, productCode);
+        product.isAlert = !product.isAlert;
+        await this.trackingProductRepository.save(product);
+    }
+
+    async getProductCurrentData(productId: string) {
+        const cacheData = await this.redis.get(`product:${productId}`);
+        if (cacheData) {
+            const { price, lowestPrice } = JSON.parse(cacheData);
+            return { price, lowestPrice };
+        }
+        const latestData = await this.productPriceModel
+            .aggregate([
+                {
+                    $match: { productId: productId },
+                },
+                {
+                    $sort: { time: -1 },
+                },
+                {
+                    $group: {
+                        _id: '$productId',
+                        price: { $first: '$price' },
+                        isSoldOut: { $first: '$isSoldOut' },
+                        lowestPrice: { $min: '$price' },
+                    },
+                },
+            ])
+            .exec();
+        const { price, lowestPrice } = latestData[0];
+        return { price, lowestPrice };
     }
 }
